removing wait class, it's part of logger now
